requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
cloudscraper>=1.2.0
requests>=2.25.0
openpyxl>=3.0.0
pathlib2>=2.3.0 